+++ examples/mnist_stages/runner.py
@@ -25,17 +25,24 @@ class IRunnerMixin(IRunner):
                 utils.set_requires_grad(layer, requires_grad=False)
         return model
 
+    def get_datasets(
+        self, stage: str, num_samples_per_class: int = None
+    ) -> "OrderedDict[str, Dataset]":
+        """Provides train/validation datasets from MNIST dataset."""
+        num_samples_per_class = num_samples_per_class or 320
+
+        datasets = super().get_datasets(stage=stage)
+        datasets["train"] = {
+            "dataset": datasets["train"],
+            "sampler": BalanceClassSampler(
+                labels=datasets["train"].targets, mode=num_samples_per_class
+            ),
+        }
+        return datasets
+
 
 class CustomSupervisedConfigRunner(IRunnerMixin, SupervisedConfigRunner):
-    def _get_dataset_from_params(self, num_samples_per_class=320, **kwargs):
-        dataset = super()._get_dataset_from_params(**kwargs)
-        if kwargs.get("train", True):
-            dataset = {
-                "dataset": dataset,
-                "sampler": BalanceClassSampler(labels=dataset.targets, mode=num_samples_per_class),
-            }
-
-        return dataset
+    pass
 
 
 if SETTINGS.hydra_required:
@@ -44,19 +51,7 @@ if SETTINGS.hydra_required:
     from catalyst.dl import SupervisedHydraRunner
 
     class CustomSupervisedHydraRunner(IRunnerMixin, SupervisedHydraRunner):
-        def _get_dataset_from_params(self, params):
-            num_samples_per_class = params.pop("num_samples_per_class", 320)
-
-            dataset = super()._get_dataset_from_params(params)
-            if params["train"]:
-                dataset = {
-                    "dataset": dataset,
-                    "sampler": BalanceClassSampler(
-                        labels=dataset.targets, mode=num_samples_per_class
-                    ),
-                }
-
-            return dataset
+        pass
 
     __all__ = ["CustomSupervisedConfigRunner", "CustomSupervisedHydraRunner"]
 else: