+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - CIFAR10 and CIFAR100 datasets from torchvision (no cv-based requirements)
   - [Catalyst Engines demo](https://github.com/catalyst-team/catalyst/tree/master/examples/engines)
   - albumentations integration ([#1238](https://github.com/catalyst-team/catalyst/pull/1238))
+  - Profiler callback ([#1226](https://github.com/catalyst-team/catalyst/pull/1226))
 
 ### Changed