+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - transform from params support for config API added ([#1236](https://github.com/catalyst-team/catalyst/pull/1236))
 - samplers from params support for config API added ([#1240](https://github.com/catalyst-team/catalyst/pull/1240))
 - recursive registry.get_from_params added ([#1241](https://github.com/catalyst-team/catalyst/pull/1241))
+- albumentations integration ([#1238](https://github.com/catalyst-team/catalyst/pull/1238))
+- Profiler callback ([#1226](https://github.com/catalyst-team/catalyst/pull/1226))
 
 ### Changed